requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
gunicorn==21.2.0
psycopg2-binary>=2.9.9
python-dotenv==1.0.0
requests==2.31.0
urllib3==2.1.0
aiohttp==3.9.1
asyncio==3.4.3
websocket-client>=1.6.0
pytz==2024.1

beautifulsoup4==4.13.4
certifi==2025.6.15
charset-normalizer==3.4.2
idna==3.10
pyfiglet==1.0.3
soupsieve==2.7
typing_extensions==4.14.1
numpy==2.3.1
